refactor(mastracode): extract truncateAnsi to shared ansi.ts module
Previously truncateAnsi lived in tool-execution-enhanced.ts and was
consumed by shell-output.ts via an import into a heavyweight module
full of unrelated tool-rendering logic. Move it to a dedicated
src/tui/components/ansi.ts so shell-output, tool-execution-enhanced,
and any future consumer can depend on a small, focused module.

Also move the truncateAnsi unit tests into their own ansi.test.ts
alongside the new module.
